+++ receiver_neo_pixel/wirelessDMX.h
@@ -65,8 +65,8 @@ typedef struct {
 DiagStatus diagStatus;
 
 typedef struct {
-  uint8_t mode;
-  uint8_t diagCode;
+  int mode;
+  int diagCode;
 } DiagBuff;
 DiagBuff diagBuff;
